Add system log entries
These are technical blog entries that I have been keeping for almost a
year now. I haven't gotten around to publishing them, and I'll proofread
them some more before publication for good. But I want them in this
repository for now so as to (a) keep them in version control, and (b)
test having files in the src/ directory that aren't being directly
copied to the server.