Explicitly copy files from src/ to site/ instead of using rsync
This change will lead to more complex compilation of certain files
(SASS, Markdown) being possible. At the moment, though, the current way
of just copying source files is still used.

This has the added advantage that each output file now correctly depends
on a particular source file, meaning that the entire site doesn't have
to be re-copied when a single file is changed.

A makefile to copy files is generated from a new perl script, and this
makefile is ran recursively. I may have been able to find a way to do
this in GNU make without relying on perl, but I wanted to keep the
Makefile POSIX compliant. Is it foolish to introduce a build dependency
on perl, just to avoid GNU make, which is just as widely available?
Perhaps. I may re-evaluate this at a later time, if I find a more
concise way of doing this in GNU make.