Add trailing slash to $(BUILDDIR) in deploy target
The intended behavior for rsync is to overwrite the remote directory
$(deploy):$(REMOTEPATH) with the local directory $(BUILDDIR) (though not
deleting remote files not present). But if there is no slash at the end of the
directory provided, then instead rsync will copy the local directory *into* the
remote directory. Previously this wasn't noticed because I had declared
$(BUILDDIR) to have a slash at the end of it. But it is clearer to have
this behavior explicit in the invocation of rsync.