Replace network links with local links
This makes it easier to test changes to the website holistically, as
clicking on links will not take me out to the live version of the site.

As I mentioned in 19a9d3f, I had trouble testing pages that used local
absolute paths because Firefox tried to load CSS files from my
computer's root directory. However, I have found that the excellent tool
http-server[0] solves this problem quite effectively. It is exactly what
I needed and is dead simple to use.

[0]: https://github.com/http-party/http-server