Remove album class
Instead, margin-top: 10px and margin-bottom: 10px is given to all
images. This doesn't seem problematic on non-album images, and it will
make it easier to write blog posts with "albums" in Markdown, because
Markdown parsers want to insert a <p> around every image/group of image,
even with an inline div. This means I would have to re-write the CSS as

    div.album > p > img {
        ...
    }

which is too hacky for my liking.