Replace copies mechanism with static/ directory
This is a lot simpler in many ways: simpler to add files to, simpler to
build, simpler to understand...

Now, files in src/ are files that must be processed in some way. Right
now the .scss files are compiled, but the .md blog entries and such are
not, yet. I have static pre-built .html files for them at the moment.