macros: Complete the program
Macros work and can be defined.

macro is in a separate module for better organization, and to make it
easier to add new macros as time goes on.

Nested macros are supported, but can be buggy, given the ad-hoc regex
matching.

Defining a macro with a regex that will always match, like `(.*)`, will
crash the program (infinite loop in process()).